jerry: add explicit casts to enum-to-integer conversions
Picked from upstream https://github.com/jerryscript-project/jerryscript/pull/2467

+++ deps/jerry/jerry-core/debugger/debugger.c
@@ -771,7 +771,7 @@ jerry_debugger_send_data (jerry_debugger_header_type_t type, /**< message type *
 
   JERRY_DEBUGGER_INIT_SEND_MESSAGE (message_type_p);
   JERRY_DEBUGGER_SET_SEND_MESSAGE_SIZE (message_type_p, 1 + size);
-  message_type_p->type = type;
+  message_type_p->type = (uint8_t) type;
   memcpy (message_type_p + 1, data, size);
 
   jerry_debugger_send (sizeof (jerry_debugger_send_type_t) + size);
